# Optimize `modified_attributes_for_create` 6-7x