I noted that after upgrading one of our apps that uses Mongoid and Mongoid-history extensively that the test suite took 33% longer to run. I tracked the slow down back to [this pull request](https://github.com/mongoid/mongoid-history/pull/158/files).

```
Warming up --------------------------------------
                HEAD     3.916k i/100ms
              v0.8.2   667.000  i/100ms
              v0.5.0    13.023k i/100ms
Calculating -------------------------------------
                HEAD     39.971k (± 3.9%) i/s -    199.716k in   5.004097s
              v0.8.2      6.580k (± 2.6%) i/s -     33.350k in   5.071545s
              v0.5.0    132.731k (± 7.3%) i/s -    664.173k in   5.031036s

Comparison:
              v0.5.0:   132731.2 i/s
                HEAD:    39971.1 i/s - 3.32x  slower
              v0.8.2:     6580.3 i/s - 20.17x  slower
```

The method is not as fast as it was in v0.5.0 because `changes` is much,
much slower than calling `attributes`.

1. The bulk of the savings comes from calling `changes` once instead of
   multiple times.
2. Some additional savings comes from not calling both `changes` and
   `attributes` and also not supporting tests that return an incorrect
   `changes` data format.